Fix pointer-width codegen in mt_evtMisc cluster

+++ src/Event/mt_evtMisc.c
@@ -724,14 +724,14 @@ void FUN_0038b140(long param_1);
 void FUN_0038b2c0(u16 *param_1,u32 *param_2,u32 *param_3,u32 *param_4, long param_5);
 void FUN_0038b550(int param_1,u32 *param_2,u32 *param_3,u32 *param_4);
 void FUN_0038b600(u32 *param_1,u32 *param_2);
-void FUN_0038b6e0(u64 param_1,int param_2,int param_3);
+void FUN_0038b6e0(int param_1,int param_2,int param_3);
 void FUN_0038b960(int param_1,long param_2,short param_3);
 void FUN_0038bc70(int param_1,short param_2,int param_3);
 void FUN_0038c460(long param_1,int param_2,int *param_3,u32 *param_4);
 void FUN_0038c540(int param_1,u64 param_2,u64 param_3,u64 param_4, u32 *param_5,u32 *param_6);
-void FUN_0038c830(u64 param_1,long param_2);
+void FUN_0038c830(int param_1,long param_2);
 void FUN_0038ca00(void);
-void FUN_0038ca80(u64 param_1);
+void FUN_0038ca80(int param_1);
 void FUN_0038cc10(int param_1,u16 *param_2,long param_3);
 void FUN_0038d560(long param_1,long param_2,u32 param_3,u32 *param_4, u32 *param_5,u32 *param_6);
 void FUN_0038d6b0(int param_1,u32 param_2);
@@ -5013,7 +5013,7 @@ void FUN_0038b600(u32 *param_1,u32 *param_2)
 // FUN_0038B6E0 NONMATCHING
 
 
-void FUN_0038b6e0(u64 param_1,int param_2,int param_3)
+void FUN_0038b6e0(int param_1,int param_2,int param_3)
 
 
 
@@ -6221,7 +6221,7 @@ void FUN_0038c540(int param_1,u64 param_2,u64 param_3,u64 param_4,
 // FUN_0038C830 NONMATCHING
 
 
-void FUN_0038c830(u64 param_1,long param_2)
+void FUN_0038c830(int param_1,long param_2)
 
 
 
@@ -6411,7 +6411,7 @@ void FUN_0038ca00(void)
 // FUN_0038CA80 NONMATCHING
 
 
-void FUN_0038ca80(u64 param_1)
+void FUN_0038ca80(int param_1)
